ZeroMemset for each Host backends
