+++ core/src/decl/Kokkos_Declare_THREADS.hpp
@@ -30,6 +30,7 @@
 #include <Threads/Kokkos_Threads_ParallelScan_Range.hpp>
 #include <Threads/Kokkos_Threads_Team.hpp>
 #include <Threads/Kokkos_Threads_UniqueToken.hpp>
+#include <Threads/Kokkos_Threads_ZeroMemset.hpp>
 #endif
 
 #endif
